[core] Fix LDAP tests

+++ apps/useradmin/src/useradmin/test_ldap.py
@@ -683,19 +683,22 @@ def test_ensure_home_directory_add_ldap_users():
     assert_false(cluster.fs.exists('/user/moe'))
 
     # Try wild card now
-    response = c.post(URL, dict(server='nonsense', username_pattern='*r*', password1='test', password2='test', ensure_home_directory=True))
+    response = c.post(URL, dict(server='nonsense', username_pattern='*rr*', password1='test', password2='test', ensure_home_directory=True))
     assert_true('/useradmin/users' in response['Location'])
     assert_true(cluster.fs.exists('/user/curly'))
     assert_true(cluster.fs.exists(u'/user/lårry'))
-    assert_true(cluster.fs.exists('/user/otherguy'))
-
-    # Clean up
-    cluster.fs.rmtree('/user/curly')
-    cluster.fs.rmtree(u'/user/lårry')
-    cluster.fs.rmtree('/user/otherguy')
+    assert_false(cluster.fs.exists('/user/otherguy'))
   finally:
+    # Clean up
     for finish in reset:
       finish()
+
+    if cluster.fs.exists('/user/curly'):
+      cluster.fs.rmtree('/user/curly')
+    if cluster.fs.exists(u'/user/lårry'):
+      cluster.fs.rmtree(u'/user/lårry')
+    if cluster.fs.exists('/user/otherguy'):
+      cluster.fs.rmtree('/user/otherguy')
 
 @attr('requires_hadoop')
 def test_ensure_home_directory_sync_ldap_users_groups():
